Day 19/requirements.txt
fastapi
uvicorn
websockets
python-dotenv
requests
google-generativeai
aiohttp
